fix(incr6): rename folder middlewares to middleware
