+++ go.mod
@@ -2,4 +2,9 @@ module github.com/Kenny201/go-yandex-shortener.git
 
 go 1.22.5
 
-require github.com/go-chi/chi/v5 v5.1.0 // indirect
+require (
+	github.com/go-chi/chi/v5 v5.1.0
+	github.com/pelletier/go-toml/v2 v2.2.2
+)
+
+require gopkg.in/yaml.v3 v3.0.1 // indirect
