+++ internal/http/middleware/logger.go
@@ -1,4 +1,4 @@
-package middlewares
+package middleware
 
 import (
 	"log/slog"
