+++ internal/http/routes.go
@@ -1,14 +1,14 @@
 package http
 
 import (
-	"github.com/Kenny201/go-yandex-shortener.git/internal/http/middlewares"
+	"github.com/Kenny201/go-yandex-shortener.git/internal/http/middleware"
 	"github.com/go-chi/chi/v5"
 )
 
 // Routes returns the initialized router
 func (s *Server) useRoutes() *chi.Mux {
 	r := chi.NewRouter()
-	r.Use(middlewares.Logger)
+	r.Use(middleware.Logger)
 
 	r.Post("/", s.handler.PostWithTextData)
 	r.Get("/{id}", s.handler.GetWithTextData)
